[None][perf] Bucket the score scratch by cohort need and widen its base offset
The score scratch was strided by max_seq_len per request*layer segment,
which at large batch was both unindexable in 32 bits and tens of GiB
(BS=256 x 8K bucket = 4.8e9 fp32 elements) for cohorts that never score
past ~1K tokens per request. The bucket now follows what cohorts
actually present (power-of-two, 1024 floor, growth reuses the existing
rebuild path), the kernel's head-plane base offset is computed in Int64,
and the setup audit guards only the remaining 32-bit plane product.

First BS=256 CuTe capture on the production nsys recipe (the geometry
both prior attempts fail-fasted on): score kernel 8.0 ms vs the C++-era
14.6 ms per round, scratch 19 GiB -> 4.8 GiB.

+++ tensorrt_llm/_torch/kv_cache_compression/triattention/triattention_cute_score.py
@@ -1138,8 +1138,14 @@ class _TriAttentionScoreKernel(_TriScoreEpilogue):
                         if cutlass.const_expr(page_half == 0):
                             cute.arch.relinquish_tmem_alloc_permit(is_two_cta=False)
                 acc_pipeline.consumer_wait(acc_consumer_state)
+                # 64-bit: the head-plane product exceeds 2^31 once the scratch
+                # spans many request*layer segments (large batch), so promote
+                # before the multiply reaches Int32 arithmetic.
                 output_offset = (
-                    kv_head * N * self.sum_seq + out_base + page_start + page_half * CTA_M
+                    cutlass.Int64(kv_head) * (N * self.sum_seq)
+                    + out_base
+                    + page_start
+                    + page_half * CTA_M
                 )
                 page_output = cute.make_tensor(
                     output.iterator + output_offset,
